+++ lib/Controller/EditorController.php
@@ -433,7 +433,7 @@ class EditorController extends Controller {
             if ($user->getUID() != $currentUserId && (!empty($email) || $operationType === "protect")) {
                 $userElement = [
                     "name" => $user->getDisplayName(),
-                    "id" => $user->getUID()
+                    "id" => $operationType === "protect" ? $this->buildUserId($user->getUID()) : $user->getUID()
                 ];
                 if (!empty($email)) {
                     $userElement["email"] = $email;
@@ -1283,7 +1283,6 @@ class EditorController extends Controller {
      * @param string $filePath - file path
      * @param string $shareToken - access token
      * @param bool $inframe - open in frame
-     * @param bool $forceEdit - open editing
      * @param bool $inviewer - open in viewer
      * @param bool $template - file is template
      * @param string $anchor - anchor for file content
@@ -1293,7 +1292,7 @@ class EditorController extends Controller {
      * @NoAdminRequired
      * @NoCSRFRequired
      */
-    public function index($fileId, $filePath = null, $shareToken = null, $inframe = false, $forceEdit = false, $inviewer = false, $template = false, $anchor = null) {
+    public function index($fileId, $filePath = null, $shareToken = null, $inframe = false, $inviewer = false, $template = false, $anchor = null) {
         $this->logger->debug("Open: $fileId $filePath ", ["app" => $this->appName]);
 
         $isLoggedIn = $this->userSession->isLoggedIn();
@@ -1332,8 +1331,7 @@ class EditorController extends Controller {
             "isTemplate" => $template,
             "inframe" => false,
             "inviewer" => $inviewer === true,
-            "anchor" => $anchor,
-            "forceEdit" => $forceEdit
+            "anchor" => $anchor
         ];
 
         $response = null;
@@ -1374,7 +1372,6 @@ class EditorController extends Controller {
      * @param integer $fileId - file identifier
      * @param string $shareToken - access token
      * @param bool $inframe - open in frame
-     * @param bool $forceEdit - open editing
      *
      * @return TemplateResponse
      *
@@ -1382,8 +1379,8 @@ class EditorController extends Controller {
      * @NoCSRFRequired
      * @PublicPage
      */
-    public function publicPage($fileId, $shareToken, $inframe = false, $forceEdit = false) {
-        return $this->index($fileId, null, $shareToken, $inframe, $forceEdit);
+    public function publicPage($fileId, $shareToken, $inframe = false) {
+        return $this->index($fileId, null, $shareToken, $inframe);
     }
 
     /**