+++ lib/TemplateManager.php
@@ -230,7 +230,7 @@ class TemplateManager {
      */
     public static function getEmptyTemplatePath($lang, $ext) {
         if (!array_key_exists($lang, self::$localPath)) {
-            $lang = "en";
+            $lang = "default";
         }
 
         return dirname(__DIR__) . DIRECTORY_SEPARATOR . "assets" . DIRECTORY_SEPARATOR . "document-templates" . DIRECTORY_SEPARATOR . self::$localPath[$lang] . DIRECTORY_SEPARATOR . "new" . $ext;
@@ -248,17 +248,21 @@ class TemplateManager {
         "cs" => "cs-CZ",
         "de" => "de-DE",
         "de_DE" => "de-DE",
+        "default" => "default",
         "el" => "el-GR",
         "en" => "en-US",
         "en_GB" => "en-GB",
         "es" => "es-ES",
         "eu" => "eu-ES",
+        "fi" => "fi-FI",
         "fr" => "fr-FR",
         "gl" => "gl-ES",
+        "he" => "he-IL",
         "it" => "it-IT",
         "ja" => "ja-JP",
         "ko" => "ko-KR",
         "lv" => "lv-LV",
+        "nb" => "nb-NO",
         "nl" => "nl-NL",
         "pl" => "pl-PL",
         "pt_BR" => "pt-BR",
@@ -266,6 +270,7 @@ class TemplateManager {
         "ru" => "ru-RU",
         "si" => "si-LK",
         "sk" => "sk-SK",
+        "sl" => "sl-SI",
         "sr" => "sr-Latn-RS",
         "sv" => "sv-SE",
         "tr" => "tr-TR",