+++ templates/editor.php
@@ -48,8 +48,7 @@ if (!empty($_["directToken"])) {
         data-template="<?php p($_["isTemplate"]) ?>"
         data-anchor="<?php p($_["anchor"]) ?>"
         data-inframe="<?php p($_["inframe"]) ?>"
-        data-inviewer="<?php p($_["inviewer"]) ?>"
-        data-forceedit="<?php p($_["forceEdit"]) ?>"></div>
+        data-inviewer="<?php p($_["inviewer"]) ?>"></div>
 
     <?php if (!empty($_["documentServerUrl"])) { ?>
         <script nonce="<?php p(\OC::$server->get(\OC\Security\CSP\ContentSecurityPolicyNonceManager::class)->getNonce()) ?>"